Save local working state for danish branch
