+++ src/index.css
@@ -70,23 +70,27 @@ body {
 
 /* ─── BUTTONS ─────────────────────────────────────────────────────────────── */
 .btn {
-  display: inline-flex; align-items: center; gap: 8px;
-  padding: 10px 20px; border-radius: var(--radius);
-  font-family: inherit; font-size: 14px; font-weight: 600;
-  cursor: pointer; border: none; transition: all 0.2s ease;
-  white-space: nowrap; text-decoration: none;
+  @apply inline-flex items-center gap-2 px-5 py-2.5 rounded-[var(--radius)] font-semibold text-sm cursor-pointer border-none transition-all whitespace-nowrap no-underline;
+  font-family: inherit;
 }
-.btn-primary { background: var(--accent); color: #fff; }
-.btn-primary:hover { transform: translateY(-1px); box-shadow: var(--glow); filter: brightness(1.1); }
-.btn-secondary { background: var(--green); color: #fff; }
-.btn-secondary:hover { filter: brightness(1.1); }
-.btn-ghost { background: transparent; color: var(--text2); border: 1px solid var(--border); }
-.btn-ghost:hover { background: var(--card); color: var(--text); }
-.btn-outline { background: transparent; color: var(--accent); border: 1px solid var(--accent); }
-.btn-outline:hover { background: rgba(0,212,255,0.1); }
-.btn-lg { padding: 14px 28px; font-size: 16px; }
-.btn-run { background: var(--green); color: #fff; font-weight: 700; }
-.btn-run:hover { filter: brightness(1.1); }
+.btn-primary { 
+  @apply bg-[var(--accent)] text-white hover:-translate-y-px hover:shadow-[var(--glow)] hover:brightness-110;
+}
+.btn-secondary { 
+  @apply bg-[var(--green)] text-white hover:brightness-110;
+}
+.btn-ghost { 
+  @apply bg-transparent text-[var(--text2)] border border-[var(--border)] hover:bg-[var(--card)] hover:text-[var(--text)];
+}
+.btn-outline { 
+  @apply bg-transparent text-[var(--accent)] border border-[var(--accent)] hover:bg-[rgba(0,212,255,0.1)];
+}
+.btn-lg { 
+  @apply px-7 py-3.5 text-base;
+}
+.btn-run { 
+  @apply bg-[var(--green)] text-white font-bold hover:brightness-110;
+}
 
 /* ─── LANDING NAV ─────────────────────────────────────────────────────────── */
 .nav {
@@ -551,9 +555,68 @@ body {
   border-radius: 6px;
 }
 
+/* ─── ZOOM TOOLBAR GLASSMORPHISM ────────────────────────────────────────────── */
+.zoom-toolbar {
+  backdrop-filter: blur(16px) saturate(1.4);
+  -webkit-backdrop-filter: blur(16px) saturate(1.4);
+  background: rgba(13, 21, 37, 0.75);
+  border: 1px solid rgba(30, 45, 71, 0.6);
+  box-shadow: 0 8px 32px rgba(0, 0, 0, 0.35), inset 0 1px 0 rgba(255, 255, 255, 0.04);
+  transition: opacity 0.2s ease, transform 0.2s ease;
+}
+[data-theme="light"] .zoom-toolbar {
+  background: rgba(248, 250, 252, 0.8);
+  border-color: rgba(203, 213, 225, 0.6);
+  box-shadow: 0 8px 32px rgba(0, 0, 0, 0.08), inset 0 1px 0 rgba(255, 255, 255, 0.6);
+}
+
+/* ─── SIMULATOR PANEL ANIMATIONS ───────────────────────────────────────────── */
+.sim-palette {
+  transition: width 0.28s cubic-bezier(0.4, 0, 0.2, 1);
+  will-change: width;
+}
+
+.sim-right-panel {
+  transition: width 0.25s cubic-bezier(0.4, 0, 0.2, 1);
+  will-change: width;
+}
+.sim-right-panel.dragging {
+  transition: none;
+}
+
+.sim-panel-content {
+  animation: panelContentIn 0.25s ease-out;
+}
+@keyframes panelContentIn {
+  from { opacity: 0; transform: translateX(12px); }
+  to   { opacity: 1; transform: translateX(0); }
+}
+
+.sim-panel-toggle {
+  transition: background 0.2s ease, box-shadow 0.2s ease, border-color 0.2s ease, left 0.2s ease;
+}
+.sim-panel-toggle:hover {
+  background: var(--bg3) !important;
+  border-color: var(--accent) !important;
+  box-shadow: 2px 0 16px rgba(0, 212, 255, 0.18), 0 0 12px rgba(0, 212, 255, 0.06);
+  color: var(--accent) !important;
+}
+
+/* ─── CANVAS MENU (animated) ───────────────────────────────────────────────── */
+.canvas-menu {
+  animation: canvasMenuIn 0.18s cubic-bezier(0.34, 1.56, 0.64, 1);
+  transform-origin: bottom right;
+}
+@keyframes canvasMenuIn {
+  from { opacity: 0; transform: scale(0.92) translateY(6px); }
+  to   { opacity: 1; transform: scale(1) translateY(0); }
+}
+
 /* ─── CANVAS MENU ITEM ──────────────────────────────────────────────────────── */
 .canvas-menu-item {
-  display: block;
+  display: flex;
+  align-items: center;
+  gap: 8px;
   width: 100%;
   text-align: left;
   background: none;
@@ -564,17 +627,17 @@ body {
   cursor: pointer;
   font-family: 'Space Grotesk', sans-serif;
   border-radius: 6px;
-  transition: background 0.13s ease, color 0.1s ease, transform 0.1s ease;
+  transition: background 0.13s ease, color 0.1s ease, transform 0.1s ease, padding-left 0.13s ease;
   outline: none;
 }
 .canvas-menu-item:hover:not([disabled]):not(.canvas-menu-item--disabled) {
-  background: var(--border);
+  background: var(--bg3);
   color: var(--text);
-  transform: translateX(3px);
+  padding-left: 16px;
 }
 .canvas-menu-item:active:not([disabled]):not(.canvas-menu-item--disabled) {
-  background: var(--border2);
-  transform: translateX(3px) scale(0.98);
+  background: var(--border);
+  transform: scale(0.985);
 }
 .canvas-menu-item--disabled {
   color: var(--text3);
