+++ src/main.jsx
@@ -7,7 +7,7 @@ import './index.css'
 const googleClientId =
   import.meta.env.VITE_GOOGLE_CLIENT_ID ||
   import.meta.env.VITE_GOOGLE_CLIENTID ||
-  ''
+  '439925019035-5qicn1624vopg9emh08dfnpu69b9qfc2.apps.googleusercontent.com'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
