+++ src/pages/ProjectGuidePage.jsx
@@ -1,7 +1,7 @@
-﻿import { useEffect, useMemo, useState } from 'react'
+import { useEffect, useMemo, useState } from 'react'
 import { useNavigate, useParams } from 'react-router-dom'
 
-const EXAMPLES_BASE_URL = 'http://localhost:5000/examples'
+const EXAMPLES_BASE_URL = import.meta.env.VITE_EXAMPLES_BASE_URL || 'http://localhost:5001/examples'
 const S = {
   page: {
     minHeight: '100vh',
