+++ src/pages/simulationpage/SimulationPage.jsx
@@ -1,0 +1,1 @@
+export { default } from './SimulatorPage.jsx'
