+++ src/services/classroomService.js
@@ -1,6 +1,6 @@
 import { getToken } from './authService.js'
 
-const BASE_URL = import.meta.env.VITE_API_BASE_URL || 'http://localhost:5000/api'
+const BASE_URL = import.meta.env.VITE_API_BASE_URL || 'http://localhost:5001/api'
 
 const authHeaders = () => {
   const token = getToken()
